requirements.txt
asgiref==3.9.1
black==25.1.0
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
distlib==0.3.9
Django==5.2.4
django-auditlog==3.2.1
django-bootstrap5==25.1
django-filter==25.1
django-tables2==2.7.5
djlint==1.36.4
EditorConfig==0.17.1
et_xmlfile==2.0.0
filelock==3.18.0
gunicorn==23.0.0
identify==2.6.12
jsbeautifier==1.15.4
json5==0.12.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
ruff==0.12.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
virtualenv==20.31.2
whitenoise==6.9.0
